requirements.txt
asgiref==3.7.1
asttokens==2.2.1
backcall==0.2.0
backports.zoneinfo==0.2.1
decorator==5.1.1
Django==4.1.7
djangorestframework==3.14.0
executing==1.2.0
ipython==8.12.2
jedi==0.18.2
matplotlib-inline==0.1.6
numpy==1.24.3
gunicorn==20.1.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
prompt-toolkit==3.0.38
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.15.1
pytz==2023.3
rdkit==2022.9.5
six==1.16.0
sqlparse==0.4.4
stack-data==0.6.2
traitlets==5.9.0
typing_extensions==4.6.2
wcwidth==0.2.6
psycopg2-binary
whitenoise
python-dotenv